Move isEditing into props
